Updated scripts to work for large data
import_NGA2 was updated to avoid looking through lists of 'OldIDs'. Rather, we split out merge events in the csv file to separate events and then create relations, then re-merge nodes together. It is a clunky work around, but it is able to run a 500k+ row csv in a few seconds rather than hours. Added timing info to outputs. Only renames the liquid core and primary drops now.

clean_csv.py was updated to prepare current nga2 outputs for this import script.

TO-DO:
- Generalize for cases without liquid core or arbitrary numbers of cores.
- Figure out breakup stage assignment.